% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bac.R
\name{create_bac}
\alias{create_bac}
\title{Creation of "PhageSim" dataset for bacteria}
\usage{
create_bac(num = 10, prop_res = 0.5)
}
\arguments{
\item{num}{The number of bacteria in the dataset}

\item{prop_res}{Starting proportion of bacteria that are resistant to antibiotics}
}
\value{
A dataset
}
\description{
Creates a dataset with a specified number of bacteria to use with other functions
}
\examples{
bac_data = create_bac(num=10, prop_res = 0.5)

}
