% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pha.R
\name{create_pha}
\alias{create_pha}
\title{Creation of "PhageSim" dataset for phage}
\usage{
create_pha(num = 10)
}
\arguments{
\item{num}{The number of phage in the dataset}
}
\value{
A dataset
}
\description{
Creates a dataset with a specified number of phage to use with other functions
}
\examples{
pha_data = create_pha(num=10)

}
