% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_any.R
\name{move_any}
\alias{move_any}
\title{Movement of bacteria or phage}
\usage{
move_any(data, size = 100, movement = 5)
}
\arguments{
\item{data}{The dataset containing the bacteria or phage information}

\item{size}{The size of the space within which to place the bacteria or phage. Must be the same as the one used to place them.}

\item{movement}{The maximum distance that bacteria or phage can move during one generation}
}
\value{
An updated dataset
}
\description{
Randomly moves bacteria or phage within the delimited space
}
\examples{
bac_data = move_any(data=bac_data, size=100, movement=5)

}
