% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_bac.R
\name{mutate_bac}
\alias{mutate_bac}
\title{Mutation of bacteria}
\usage{
mutate_bac(data, mutate_bac = 0.1)
}
\arguments{
\item{data}{The dataset containing the bacteria information}

\item{mutate_bac}{Mutation rate of bacteria}
}
\value{
An updated dataset
}
\description{
Randomly mutate bacteria, depending on mutation rate
}
\examples{
bac_data = mutate_bac(data=bac_data, mutate_bac=2)

}
