% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_any.R
\name{place_any}
\alias{place_any}
\title{Placement of bacteria or phage}
\usage{
place_any(data, size = 100)
}
\arguments{
\item{data}{The dataset containing the bacteria or phage information}

\item{size}{The size of the space within which to place the bacteria or phage}
}
\value{
An updated dataset
}
\description{
Places the bacteria or phage randomly within a delimited space
}
\examples{
bac_data = place_any(data=bac_data, size=100)

}
