% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Create the transmission kernel parameters}
\usage{
set_params(
  size = 250,
  rep_bac = 2,
  movement_bac = 10,
  movement_pha = 10,
  beta = 0.5,
  burst = 10,
  trans = 0.1,
  decay = 0.5
)
}
\arguments{
\item{size}{The size of the space within which the bacteria and phage can move.}

\item{rep_bac}{Replication rate of bacteria}

\item{movement_bac}{The maximum distance that bacteria can move during one generation}

\item{movement_pha}{The maximum distance that phage can move during one generation}

\item{beta}{Phage infection probability}

\item{burst}{Phage burst size}

\item{trans}{Probability of transduction instead of normal replication}
}
\value{
A vector containing the parameter values
}
\description{
This is needed to execute the update_all and simulate_all functions
}
\examples{
params = set_params()

}
