% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{update_all}
\alias{update_all}
\title{Update bacteria infections by phage}
\usage{
update_all(data_bac, data_pha, p, tr_count = 0)
}
\arguments{
\item{data_bac}{The dataset containing the bacteria information}

\item{data_pha}{The dataset containing the phage information}

\item{p}{General parameters. These must have been created using the set_params function.}
}
\value{
An updated dataset
}
\description{
Updates the bacteria infections by phage, generating or decaying phage
}
\examples{
dots_data = update_dots(data=dots_data, gen=0, p=params)

}
